2n

+++ pg4.c
@@ -1,12 +1,13 @@
 #include<stdio.h>
-int rev(int n)
+long int rev(long int n)
 {
 if(n==0)
 return 0;
+
 return 1+rev(n/10);
 }
 int main()
 {
-printf("%d",rev(2334));
+printf("%ld",rev(2334));
 }
